Move InstallAction.
Move InstallAction to it’s home with the other Flutter actions.
