+++ src/io/flutter/actions/InstallSdkAction.java
@@ -1,9 +1,9 @@
 /*
- * Copyright 2017 The Chromium Authors. All rights reserved.
+ * Copyright 2018 The Chromium Authors. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
  */
-package io.flutter.module;
+package io.flutter.actions;
 
 import com.intellij.execution.ExecutionException;
 import com.intellij.execution.configurations.GeneralCommandLine;
